feature: Added layout for authentication pages
