+++ config/constants.ts
@@ -1,3 +1,3 @@
-export const API_URL = process.env.NEXT_PUBLIC_API_URL
+export const API_URL = process.env.NEXT_PUBLIC_API_URL;
 export const APP_NAME = process.env.NEXT_PUBLIC_APP_NAME || "FlexTest";
-export const MOCKARO_KEY = process.env.NEXT_PUBLIC_MOCKARO_KEY
+export const MOCKARO_KEY = process.env.NEXT_PUBLIC_MOCKARO_KEY;
