+++ pages/index.ts
@@ -1,0 +1,12 @@
+import { useEffect } from "react";
+import { useRouter } from "next/router";
+
+export default function Home() {
+  const router = useRouter();
+
+  useEffect(() => {
+    router.replace("/auth/login");
+  }, [router]);
+
+  return null;
+}
