+++ tsconfig.json
@@ -34,6 +34,7 @@
       "@middleware/*": ["middleware/*"],
       "@tests/*": ["tests/*"],
       "@mocks/*": ["mocks/*"],
+      "@layouts/*": ["layouts/*"],
     }
   },
   "include": [
